+++ Model/BossRush.cs
@@ -1,0 +1,7 @@
+﻿namespace lib.remnant2.analyzer.Model;
+
+// Boss rush worlds do not have much in it, so this is the trimmed down version of RolledWorld
+public class BossRush : WorldMode
+{
+    public required LootItem Mode { get; set; }
+}